Fix usage of multiple actions to determine if user interaction is needed
We make it possible for one specific dbus method to use multiple actions
to know if user interaction is needed.s

It used to work before without doing any specific, but with the port to
PolicyKit 1, we need to specify that the first actions (until the last
one) cannot use user interaction. So if the user is authorized for them,
it will just work, else we'll check for the following actions.

User interaction is only acceptable for the last action, which is the
minimal one required.

For example: to enable a printer, the fine-grained privilege
printer-enable is enough, but if it requires user interaction, we'll
skip it and go straight to the printer-X-edit action, which makes it
possible for the user to do more changes -- we won't ask for
authentication to enable it, and then again to change an option.

https://bugzilla.redhat.com/show_bug.cgi?id=518012