Fix a bunch of issues when getting/putting a file from cups
There was basically no check for permissions. Now, we temporarily change
our effective uid/gid to the one of the user to open the file for
writing (when getting) or reading (when putting). We then only use
operations that work on the file descriptor to avoid potential race
conditions.

Before that, people could:
 - overwrite any file with the content of a cups resource
 - put any file in a cups resource

Part of fix for CVE-2012-4510.